Update README.md
The "Email Validation Using String" project is a Python program that validates the format and structure of an email address. It checks various criteria such as length, presence of "@" symbol, correct placement of domain extension, and allowed characters. The program provides feedback on the validity of the email address entered by the user, indicating whether it is correct or specifying the reason for its invalidity.